Extract GetStatusInformation as a pure helper (#18423)
'GetStatusInformation' is renamed 'GetInformation' and is now a pure
helper. This prepare the path to migration metadata as a component.
