+++ comp/metadata/host/utils/host_info_notwin_test.go
@@ -1,0 +1,27 @@
+// Unless explicitly stated otherwise all files in this repository are licensed
+// under the Apache License Version 2.0.
+// This product includes software developed at Datadog (https://www.datadoghq.com/).
+// Copyright 2016-present Datadog, Inc.
+
+//go:build !windows
+
+package utils
+
+import (
+	"testing"
+
+	"github.com/DataDog/datadog-agent/pkg/util/cache"
+	"github.com/shirou/gopsutil/v3/host"
+	"github.com/stretchr/testify/assert"
+)
+
+func TestGetHostInfo(t *testing.T) {
+	assert.NotNil(t, GetInformation())
+	fakeInfo := &host.InfoStat{HostID: "FOOBAR"}
+
+	cache.Cache.Set(cacheKey, fakeInfo, cache.NoExpiration)
+	defer cache.Cache.Delete(cacheKey)
+
+	info := GetInformation()
+	assert.Equal(t, fakeInfo, info)
+}
