+++ pkg/metadata/host/host_notwin.go
@@ -12,11 +12,10 @@ import (
 	"strings"
 
 	"github.com/shirou/gopsutil/v3/cpu"
-	"github.com/shirou/gopsutil/v3/host"
 
+	hostMetadataUtils "github.com/DataDog/datadog-agent/comp/metadata/host/utils"
 	"github.com/DataDog/datadog-agent/pkg/metadata/inventories"
 	"github.com/DataDog/datadog-agent/pkg/util/cache"
-
 	"github.com/DataDog/datadog-agent/pkg/util/log"
 )
 
@@ -24,9 +23,7 @@ import (
 func InitHostMetadata() error {
 	var err error
 	_, err = cpu.Info()
-
 	return err
-
 }
 
 func getSystemStats() *systemStats {
@@ -36,8 +33,6 @@ func getSystemStats() *systemStats {
 		stats = x.(*systemStats)
 	} else {
 		cpuInfo := getCPUInfo()
-		hostInfo := getHostInfo()
-
 		stats = &systemStats{
 			Machine:   runtime.GOARCH,
 			Platform:  runtime.GOOS,
@@ -47,6 +42,7 @@ func getSystemStats() *systemStats {
 		}
 
 		// fill the platform dependent bits of info
+		hostInfo := hostMetadataUtils.GetInformation()
 		fillOsVersion(stats, hostInfo)
 		cache.Cache.Set(key, stats, cache.NoExpiration)
 		inventories.SetHostMetadata(inventories.HostOSVersion, getOSVersion(hostInfo))
@@ -71,25 +67,4 @@ func getCPUInfo() *cpu.InfoStat {
 	info := &i[0]
 	cache.Cache.Set(key, info, cache.NoExpiration)
 	return info
-}
-
-func getHostInfo() *host.InfoStat {
-	key := buildKey("hostInfo")
-	if x, found := cache.Cache.Get(key); found {
-		return x.(*host.InfoStat)
-	}
-
-	info, err := host.Info()
-	if err != nil {
-		// don't cache and return zero value
-		log.Errorf("failed to retrieve host info: %s", err)
-		return &host.InfoStat{}
-	}
-	cache.Cache.Set(key, info, cache.NoExpiration)
-	return info
-}
-
-// GetStatusInformation just returns an InfoStat object, we need some additional information that's not
-func GetStatusInformation() *host.InfoStat {
-	return getHostInfo()
 }
